refactor: migration filename correction
