+++ packages/nocodb/src/lib/noco/common/XcMigrationSourcev2.ts
@@ -1,7 +1,7 @@
 import * as nc_011 from '../migrationsv2/nc_011';
 import * as nc_012_alter_column_data_types from '../migrationsv2/nc_012_alter_column_data_types';
-import * as nc_013_alter_column_data_types from '../migrationsv2/nc_013_alter_column_data_types';
 import * as nc_013_sync_source from '../migrationsv2/nc_013_sync_source';
+import * as nc_014_alter_column_data_types from '../migrationsv2/nc_014_alter_column_data_types';
 
 // Create a custom migration source class
 export default class XcMigrationSourcev2 {
@@ -14,7 +14,7 @@ export default class XcMigrationSourcev2 {
       'nc_011',
       'nc_012_alter_column_data_types',
       'nc_013_sync_source',
-      'nc_013_alter_column_data_types'
+      'nc_014_alter_column_data_types'
     ]);
   }
 
@@ -30,8 +30,8 @@ export default class XcMigrationSourcev2 {
         return nc_012_alter_column_data_types;
       case 'nc_013_sync_source':
         return nc_013_sync_source;
-      case 'nc_013_alter_column_data_types':
-        return nc_013_alter_column_data_types;
+      case 'nc_014_alter_column_data_types':
+        return nc_014_alter_column_data_types;
     }
   }
 }
